dtprobed: fix probe name debug output
Debug output while reading dof_parsed_t structures was printing the
probe names as provider->provider.name and probe->probe.name, but
probe.name is a 0-separated concatenation of the probe name elements.
It only printed the module name.  Corrected to print the full probe
name.

Also moved it to the proper place (when a new probe is read rather
than for every tracepoint).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>