selftests/usdt: add is-enabled stapsdt tests using semaphores
Is-eanbled probes are implemented using semaphores where the
semaphore address is specified in the ELF notes and passed
in at probe creation time to have the kernel reference-count
probes; this allows us to have argument assembly code that
only gets executed when the stapsdt probe is in use.

Test is-enabled probes and ensure that the kernel-controlled
semaphore is not incremented for disabled probes.  Also test
that provider/probe names with double-underscores are converted
to dashes, i.e. one__probe because :one-probe.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>