test: Fix post processing of return-action override-pid tests
In commit ee8c066e8 ("rawfbt: selectively allow return() in clauses"),
tests were introduced to check that return values are overwritten.  A
pid check value is output and used to facilitate comparisons to a .r
results file.

The post processing logic was flawed in two respects:

1 The check value was being pulled from the "01" entry.  Therefore,
  the "00" entry, typically appearing first, would not be converted.

2 Values were not sorted.  Sometimes this means the disorder just
  described would be avoided and the test would pass.  Mostly,
  it means that behavior can be erratic.

Sort results and use the "00" pid value for converting test results.
Add $dt_flags to the $dtrace line while we're at it.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>