Refactor the versioning handling system (consolidate data)
Rather than having multiple data items that need to be updated whenever
a new version is released, a single entry in versions.list will now be
all that is needed.  The mkvers script generates the necessary defines,
and is also used to return the most recent release version to be used
in GNUmakefile.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>