examples: Add "language features" demo scripts
The idea is to have lots of short scripts, each illustrating a
D language feature.

The numbering scheme serves two purposes:

*)  the scripts are more or less in ascending order
    of difficulty, although that's admittedly a tricky
    concept in such a multi-dimensional space

*)  there is some grouping of themes -- e.g.,
    "introduction", "providers", "built-in variables", etc.

The combination of comments and code is intended to help either
a human user or an AI bot to learn how to use the D language.

Other examples have been based on functional objectives (doing
simple but useful tasks).  These new examples are based on language
features.  A future set of examples could be based on errors!  That
is, the scripts generate errors of different sorts that we would
illustrate (and explain how to fix).

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Ruud van der Pas <ruud.vanderpas@oracle.com>