Convert ip6_finish_output to rawfbt
With commit fe2101e55 ("fbt: implement based on fentry/fexit probes")
test/unittest/ip/tst.ipv6localicmp.sh started to fail.  The problem
appears to be that fbt::ip6_finish_output:entry, which is needed for
ip:::send, does not fire.

Use rawfbt for this probe to revert back to kprobes in this case.

Other fbt probes may also be impacted, but this case showed up in testing.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>