printf: fake type node is only needed for aggregation printing
Regular printf validation was passing DTRACEACT_AGGREGATION as kind
argument, even though the action was not doing anything with
aggregations.  Pass DTRACEACT_NONE instead because the argument is not
used in that case.

Also, do not populate the fake type node if the validation is called
for anything other than aggregations.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>