enum: support declarations with a trailing comma
When doing python tracing, it was recently observed that #include'ing
a file with enum declarations with trailing commas fails; this is
due to the D grammar being strict about the last enumerator value
not having a trailing comma.  So

enum foo {
	BAR,
	BAZ
};

is permitted, but

enum foo {
	BAR,
	BAZ,
};

is not.  The latter pattern is used quite frequently in #include
files, especially where conditional compilation of some enum
values is done.

Relax this constraint and add a test to validate that D compilation
succeeds with the trailing comma in an enum declaration.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>