test: Fix some ustackdepth tests
Fix these issues.

Issue 1.

In commit e48bf6a23 ("Add support for built-in variable ustackdepth"),
tests ustackdepth2 and ustackdepth_big were introduced and their .r and
.r.p files were links to those of the stackdepth2 test.

In commit 842eabc0e
("test: caller and stackdepth tests for fbt provider"), the stackdepth2
test was explicitly renamed "fbt" since tests for many providers were
being introduced.  The links of the earlier patch were not updated.

Issue 2.

The ustackdepth_big test modified perf_event_max_stack.  If the test
were interrupted, the value was not restored.

Issue 3.

That ustackdepth_big test only checked that the stackdepth was
consistent with the reported stack.  While this check is good, it missed
the point of the test, which is that the system limit could be
increased.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>