doc: Link to "Character Escape Sequences" table
A link to this table did not work since it referenced an ID that did not
exist.  Add an anchor ID by the table and use it.

While we're at it, convert the table from HTML to Markdown and break up
some longer lines, improving readability and maintainability of the
Markdown source file.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Elena Zannoni <elena.zannoni@oracle.com>