doc: Add include path for sys/usdt.h
When dtrace constructs a USDT header, it includes the header sys/usdt.h.

Add a note to our documentation that /usr/lib64/dtrace/include needs to
be on the include path for this header to be found.

While we are at it, break long lines of text up for better readability
of the Markdown file and simpler diffs in the future.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>